✨ post 를 불러오는 함수를 컴포저블로 분리
- markdown utils 함수로 분리
